+++ fields/types/select/SelectField.js
@@ -26,30 +26,44 @@ module.exports = Field.create({
 		});
 	},
 
+	/*
+	** Convert label to string if it is multilingual object
+	** Terry Chan
+	** 27/11/2018
+	*/
+	convertObject (label) {
+		const { currentLang } = this.props;
+		if (currentLang && typeof label === 'object') {
+			return label[currentLang];
+		} else if (typeof label !== 'object') {
+			return label;
+		}
+		return null;
+	},
+
 	renderValue () {
 		const { ops, value } = this.props;
 		const selected = ops.find(opt => opt.value === value);
-
 		return (
 			<FormInput noedit>
-				{selected ? selected.label : null}
+				{selected ? this.convertObject(selected.label) : null}
 			</FormInput>
 		);
 	},
 
 	renderField () {
-		const { numeric, ops, path, value: val } = this.props;
-
+		const { numeric, boolean, ops, path, value: val, currentLang } = this.props;
 		// TODO: This should be natively handled by the Select component
-		const options = (numeric)
-			? ops.map(function (i) {
-				return { label: i.label, value: String(i.value) };
-			})
-			: ops;
-		const value = (typeof val === 'number')
-			? String(val)
-			: val;
-
+		var options = ops;
+		if (ops.length) {
+			options = ops.map(function (i) {
+				return { label: this.convertObject(i.label), value: String(i.value) };
+			}, this);
+		}
+		const value = String(val);
+		// if (typeof val === 'number' || typeof val === 'boolean') {
+		// 	value = String(val)
+		// }
 		return (
 			<div>
 				{/* This input element fools Safari's autocorrect in certain situations that completely break react-select */}